requirements.txt
fastapi==0.122.0
uvicorn[standard]==0.41.0
sqlalchemy==2.0.44
pydantic==2.12.4
python-dotenv==1.2.1
psycopg2-binary==2.9.11
httpx==0.28.1
